test: reorganize widget navigation suites
