+++ tests/test_widgets_navigation.py
@@ -1,6 +1,8 @@
-from tests.widgets_navigation.test_navigation_lists import *  # noqa: F403
-from tests.widgets_navigation.test_track_list_widget import *  # noqa: F403
-from tests.widgets_navigation.test_lyrics_paste_behavior import *  # noqa: F403
-from tests.widgets_navigation.test_main_window_instrumental import *  # noqa: F403
-from tests.widgets_navigation.test_navigation_controller import *  # noqa: F403
+from tests.widgets_navigation.navigation.test_navigation_lists import *  # noqa: F403
+from tests.widgets_navigation.track_list.test_track_list_widget import *  # noqa: F403
+from tests.widgets_navigation.lyrics.test_lyrics_editor_widget import *  # noqa: F403
+from tests.widgets_navigation.main_window.test_main_window import *  # noqa: F403
+from tests.widgets_navigation.navigation.test_navigation_controller import *  # noqa: F403
+from tests.widgets_navigation.settings.test_settings_dialog import *  # noqa: F403
+from tests.widgets_navigation.shared_ui.test_ui_components import *  # noqa: F403
