+++ tests/widgets_navigation/navigation/test_navigation_controller.py
@@ -1,5 +1,6 @@
 from tests.widgets_navigation._shared import *
 
+
 @unittest.skipUnless(HAS_QT, "PySide6 is required for widget tests")
 class NavigationControllerTests(unittest.TestCase):
     @classmethod
